coupon update

+++ src/coupons/coupons/coupons.controller.ts
@@ -1,23 +1,53 @@
 import {
-  Controller,
-  Get,
-  Post,
+  BadRequestException,
   Body,
-  Param,
+  Controller,
   Delete,
+  Get,
+  Headers,
+  Param,
+  Post,
   Put,
   Query,
+  UseGuards,
 } from '@nestjs/common';
-import { CouponsService } from './coupons.service';
+import { InjectRepository } from '@nestjs/typeorm';
+import * as jwt from 'jsonwebtoken';
+import { User } from 'src/users/entities/user.entity';
+import { AuthTokenGuard } from 'src/users/guards/authTokenGuard';
+import { Repository } from 'typeorm';
 import { CreateCouponDto } from '../dto/create-coupon.dto';
 import { UpdateCouponDto } from '../dto/update-coupon.dto';
+import { CouponsService } from './coupons.service';
 
 @Controller('coupons')
 export class CouponsController {
-  constructor(private readonly service: CouponsService) {}
+  constructor(
+    private readonly service: CouponsService,
+    @InjectRepository(User)
+    private readonly userRepository: Repository<User>,
+  ) {}
 
+  @UseGuards(AuthTokenGuard)
   @Post()
-  async create(@Body() dto: CreateCouponDto) {
+  async create(
+    @Body() dto: CreateCouponDto,
+    @Headers('user-secret') userSecret: string,
+  ) {
+    if (!userSecret) {
+      throw new BadRequestException('user-secret not found in headers');
+    }
+    const decodedUser: any = jwt.decode(userSecret);
+
+    const user = await this.userRepository.findOne({
+      where: {
+        id: decodedUser.UserId,
+      },
+    });
+
+    if (!user) {
+      throw new BadRequestException('user not found against provided token');
+    }
     return await this.service.create(dto);
   }
 
@@ -34,13 +64,52 @@ export class CouponsController {
     return await this.service.findOne(+id);
   }
 
+  @UseGuards(AuthTokenGuard)
   @Put(':id')
-  async update(@Param('id') id: string, @Body() dto: UpdateCouponDto) {
+  async update(
+    @Param('id') id: string,
+    @Headers('user-secret') userSecret: string,
+    @Body() dto: UpdateCouponDto,
+  ) {
+    if (!userSecret) {
+      throw new BadRequestException('user-secret not found in headers');
+    }
+
+    const decodedUser: any = jwt.decode(userSecret);
+
+    const user = await this.userRepository.findOne({
+      where: {
+        id: decodedUser.UserId,
+      },
+    });
+
+    if (!user) {
+      throw new BadRequestException('user not found against provided token');
+    }
     return await this.service.update(+id, dto);
   }
 
+  @UseGuards(AuthTokenGuard)
   @Delete(':id')
-  async remove(@Param('id') id: string) {
+  async remove(
+    @Param('id') id: string,
+    @Headers('user-secret') userSecret: string,
+  ) {
+    if (!userSecret) {
+      throw new BadRequestException('user-secret not found in headers');
+    }
+
+    const decodedUser: any = jwt.decode(userSecret);
+
+    const user = await this.userRepository.findOne({
+      where: {
+        id: decodedUser.UserId,
+      },
+    });
+
+    if (!user) {
+      throw new BadRequestException('user not found against provided token');
+    }
     return await this.service.remove(+id);
   }
